#184 Windows 64bit Build break fixes

+++ packager/media/base/container_names_unittest.cc
@@ -191,7 +191,7 @@ TEST(ContainerNamesTest, WebVtt) {
 
   EXPECT_EQ(CONTAINER_WEBVTT,
             DetermineContainer(webvtt_with_utf8_byte_order_mark.data(),
-                               webvtt_with_utf8_byte_order_mark.size()));
+              base::checked_cast<int>(webvtt_with_utf8_byte_order_mark.size())));
 }
 
 TEST(ContainerNamesTest, FileCheckOGG) {
