+++ packager/media/formats/mp2t/es_parser_h26x_unittest.cc
@@ -215,7 +215,7 @@ void EsParserH26xTest::RunTest(Nalu::CodecType codec_type,
           seen_key_frame = true;
       }
 
-      std::vector<uint8_t> es_data = CreateNalu(codec_type, types[k], k);
+      std::vector<uint8_t> es_data = CreateNalu(codec_type, types[k], base::checked_cast<int>(k));
       cur_sample_data.push_back(0);
       cur_sample_data.push_back(0);
       cur_sample_data.push_back(0);
@@ -235,7 +235,7 @@ void EsParserH26xTest::RunTest(Nalu::CodecType codec_type,
       while (offset < es_data.size()) {
         // Insert the data in parts to test partial data searches.
         size = std::min(size + 1, es_data.size() - offset);
-        ASSERT_TRUE(es_parser.Parse(&es_data[offset], size, pts, dts));
+        ASSERT_TRUE(es_parser.Parse(&es_data[offset], base::checked_cast<int>(size), pts, dts));
         offset += size;
       }
     }
